Moves some common content into helper
